requirements.txt
langchain
openai
python-dotenv
chromadb
unstructured
pdfminer.six